+++ src/th_hash.h
@@ -11,10 +11,10 @@
  * @brief Fowler-Noll-Vo hash function (FNV-1a).
  * See https://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function
  */
-TH_INLINE(uint32_t)
+TH_INLINE(size_t)
 th_hash_bytes(const void* data, size_t len)
 {
-    uint32_t hash = 2166136261u;
+    size_t hash = 2166136261u;
     const uint8_t* bytes = (const uint8_t*)data;
     for (size_t i = 0; i < len; ++i) {
         hash ^= bytes[i];
@@ -23,7 +23,7 @@ th_hash_bytes(const void* data, size_t len)
     return hash;
 }
 
-TH_INLINE(uint32_t)
+TH_INLINE(size_t)
 th_hash_cstr(const char* str)
 {
     return th_hash_bytes(str, strlen(str));
